proc_comm_wince: unset and handle all pending interrupt bits together
We've only been handling and unsetting the first set interrupt bit we
found. In cases there was more then one bit set, there was no subsequent
notification of ARM9 telling us to handle the remaing ones.

We now unset all pending interrupt bits and try to handle all in the
worker as far as we know them.

This seems to fix the random SODs just recently appearing.

--

Use disable_irq_nosync() instead of writing into the VIC register
directly. disable_irq() works on .27 but on .35 likely due a change
of implementation that won't work anymore. It tries to sync all
interrupt handlers - even ourself - which apparently fails as we
are currently just within the isr. (Vibrate loop on haret boot.)